event_flow/service/service/backend/requirements.txt
flask
waitress
mysql-connector-python
PyJWT
flask-cors
bcrypt